lambda/requirements.txt
ask-sdk-core==1.11.0
google-genai>=1.33.0